Data.Type.Lens

Description

Lenses and optics for manipulating DataKind-based types, powered by singletons defunctionalization.

See Data.Type.Lens.Examples for example usage and syntax.

For the most part, you should be able to use them just like you'd use the functions from the lens or microlens libraries; just remember to capitalize names like Over and Set, since they are type families.

Note that the ways of "creating" a lens or optic (Sets_, Traverse_, To_, MkLens_, etc. are all suffixed with _ for convenience, to reserve the underscoreless identifiers for the fully applied type family as per singletons library convention.

Documentation

type LensLike f s t a b = (a -> f b) -> s -> f t

The general shape of optics in this library. ("van Laarhoven")

For different levels of polymorphism on f, you get different types of optics:

If f can be any Functor, you have a Lens (see ALens)

If f is only Identity, you have a setter (see ASetter)

If f is only Const R for a specific R, you have a getter of R (see Getting)

If f can be Const r for any Monoid r, you have a Fold.

If f can be any Applicative, you have a Traversal (see ATraversal)

Normal lens libraries implement the constraints for lenses, folds, and traversals using RankN types, but we don't do that here to avoid working with RankN kinds.

type LensLike' f s a = LensLike f s s a a

A LensLike that does not change any types.

Setting

type ASetter s t a b = LensLike Identity s t a b

A settable "lens". Usable with Over (%~), constructable with To or any of the general lens constructors.

See LensLike for more information.

Using

Ways of consuming a setter.

type family Over (a :: (~>) ((~>) a (Identity b)) ((~>) s (Identity t))) (a :: (~>) a b) (a :: s) :: t where ...

Equations

Over l f x = Case_6989586621679207752 l f x (Let6989586621679207748Scrutinee_6989586621679203526Sym3 l f x)

type (%~) l f = OverSym2 l f infixr 4

Infix application of Over

sOver :: forall s t a b (t :: (~>) ((~>) a (Identity b)) ((~>) s (Identity t))) (t :: (~>) a b) (t :: s). Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply OverSym0 t) t) t :: t)

type family Set (a :: (~>) ((~>) a (Identity b)) ((~>) s (Identity t))) (a :: b) (a :: s) :: t where ...

Equations

Set l y a_6989586621679207762 = Apply (Apply (Apply OverSym0 l) (Apply (Apply (Apply Lambda_6989586621679207767Sym0 l) y) a_6989586621679207762)) a_6989586621679207762

type (.~) l x = SetSym2 l x infixr 4

Infix application of Set

sSet :: forall s t a b (t :: (~>) ((~>) a (Identity b)) ((~>) s (Identity t))) (t :: b) (t :: s). Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply SetSym0 t) t) t :: t)

Making

Ways of creating a setter-only.

type Sets_ f = SetsSym1 f

Create a Setter from a setting function.

Sets_ :: ((a ~> b) ~> (s ~> t)) -> ASetter s t a b

type family Sets (a :: (~>) ((~>) a b) ((~>) s t)) (a :: (~>) a (Identity b)) (a :: s) :: Identity t where ...

Equations

Sets f g a_6989586621679207704 = Apply (Apply (Apply (.@#@$) IdentitySym0) (Apply f (Apply (Apply (Apply Lambda_6989586621679207709Sym0 f) g) a_6989586621679207704))) a_6989586621679207704

sSets :: forall a b s t (t :: (~>) ((~>) a b) ((~>) s t)) (t :: (~>) a (Identity b)) (t :: s). Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply SetsSym0 t) t) t :: Identity t)

Getting

type Getting r s a = LensLike (Const r) s s a a

A retrieving "lens". If r is fixed to a type, it's a Getter for that type. If r is polymorphic over all Monoid, then it's a Fold over as.

As a Getter, usable with View (^.); as a Fold, usable with ToListOf (^..), Preview (^?), etc.

Normal lens libraries implement the constraints for folds using RankN types, but we don't do that here to avoid working with RankN kinds.

See LensLike for more information.

Using

Ways of consuming a getter

type family View (a :: (~>) ((~>) a ((Const a :: Type -> Type) a)) ((~>) s ((Const a :: Type -> Type) s))) (a :: s) :: a where ...

Equations

View l x = Case_6989586621679207735 l x (Let6989586621679207732Scrutinee_6989586621679203531Sym2 l x)

type (^.) x l = View l x infixl 8

Infix application of View

sView :: forall a s (t :: (~>) ((~>) a ((Const a :: Type -> Type) a)) ((~>) s ((Const a :: Type -> Type) s))) (t :: s). Sing t -> Sing t -> Sing (Apply (Apply ViewSym0 t) t :: a)

Making

Ways of creating a getter-only.

type To_ f = ToSym1 f

Create a Getter from a getting function.

To_ :: (a ~> b) -> Getting b a b

type family To (a :: (~>) s a) (a :: (~>) a ((Const r :: Type -> Type) a)) (a :: s) :: (Const r :: Type -> Type) s where ...

Equations

To f g x = Case_6989586621679207694 f g x (Let6989586621679207690Scrutinee_6989586621679203539Sym3 f g x)

sTo :: forall s a r (t :: (~>) s a) (t :: (~>) a ((Const r :: Type -> Type) a)) (t :: s). Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply ToSym0 t) t) t :: (Const r :: Type -> Type) s)

Lenses

type ALens s t a b = LensLike (Context a b) s t a b

If a function expects an ALens, it can be given any Lens (a LensLike f that works for any Functor f).

You can use an ALens as a normal lens by using CloneLens_.

Making

Ways of creating a lens

type MkLens_ f g = MkLensSym2 f g

Create a Lens from a setter and a getter.

MkLens_
    :: Functor f
    => (s ~> a)
    -> (s ~> b ~> t)
    -> LensLike f s t a b

type family MkLens (a :: (~>) s a) (a :: (~>) s ((~>) b t)) (a :: (~>) a (f b)) (a :: s) :: f t where ...

Equations

MkLens v s f x = Apply (Apply (<$>@#@$) (Apply s x)) (Apply f (Apply v x))

sMkLens :: forall f s a b t (t :: (~>) s a) (t :: (~>) s ((~>) b t)) (t :: (~>) a (f b)) (t :: s). SFunctor f => Sing t -> Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply (Apply MkLensSym0 t) t) t) t :: f t)

Cloning

type CloneLens_ l = CloneLensSym1 l

Clone a polymorphic lens so it can be used as more than one type of thing (getter or setter).

CloneLens_
    :: Functor f
    => LensLike (Context a b) s t a b
    -> LensLike f s t a b

Useful for writing a function that takes a lens and uses it in more than one way; if you have it take an ALens, you can then use CloneLens_ to use it as a getter or setter.

type family CloneLens (a :: (~>) ((~>) a (Context a b b)) ((~>) s (Context a b t))) (a :: (~>) a (f b)) (a :: s) :: f t where ...

Equations

CloneLens l f x = Case_6989586621679207664 l f x (Let6989586621679207652Scrutinee_6989586621679203544Sym3 l f x)

sCloneLens :: forall f a b s t (t :: (~>) ((~>) a (Context a b b)) ((~>) s (Context a b t))) (t :: (~>) a (f b)) (t :: s). SFunctor f => Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply CloneLensSym0 t) t) t :: f t)

Traversals and Folds

type ATraversal s t a b = LensLike (Bazaar a b) s t a b

If a function expects an ATraversal, it can be given any Traversal (a LensLike f that works for any Applicative f).

You can use an ATraversal as a normal traversal by using CloneTraversal_.

Using

Ways of consuming traversals and folds

type family Preview (a :: (~>) ((~>) a ((Const (First a) :: Type -> Type) a)) ((~>) s ((Const (First a) :: Type -> Type) s))) (a :: s) :: Maybe a where ...

Equations

Preview l x = Case_6989586621679207606 l x (Let6989586621679207603Scrutinee_6989586621679203556Sym2 l x)

type (^?) x l = Preview l x infixl 8

Infix application of Preview

sPreview :: forall a s (t :: (~>) ((~>) a ((Const (First a) :: Type -> Type) a)) ((~>) s ((Const (First a) :: Type -> Type) s))) (t :: s). Sing t -> Sing t -> Sing (Apply (Apply PreviewSym0 t) t :: Maybe a)

type family ToListOf (a :: (~>) ((~>) a ((Const [a] :: Type -> Type) a)) ((~>) s ((Const [a] :: Type -> Type) s))) (a :: s) :: [a] where ...

Equations

ToListOf l x = Case_6989586621679207639 l x (Let6989586621679207629Scrutinee_6989586621679203549Sym2 l x)

type (^..) x l = ToListOf l x infixl 8

Infix application of ToListOf

sToListOf :: forall a s (t :: (~>) ((~>) a ((Const [a] :: Type -> Type) a)) ((~>) s ((Const [a] :: Type -> Type) s))) (t :: s). Sing t -> Sing t -> Sing (Apply (Apply ToListOfSym0 t) t :: [a])

type family UnsafePreview (a :: (~>) ((~>) a ((Const (First a) :: Type -> Type) a)) ((~>) s ((Const (First a) :: Type -> Type) s))) (a :: s) :: a where ...

Equations

UnsafePreview l x = Case_6989586621679207619 l x (Let6989586621679207616Scrutinee_6989586621679203563Sym2 l x)

type (^?!) x l = UnsafePreview l x infixl 8

Infix application of UnsafePreview

sUnsafePreview :: forall a s (t :: (~>) ((~>) a ((Const (First a) :: Type -> Type) a)) ((~>) s ((Const (First a) :: Type -> Type) s))) (t :: s). Sing t -> Sing t -> Sing (Apply (Apply UnsafePreviewSym0 t) t :: a)

Making

Ways of creating traversals and folds

type Folding_ f = FoldingSym1 f

Create a Fold from a "folding function":

Folding_
    :: (Foldable f, Monoid r)
    => (s ~> f a)
    -> Getting r s a

type family Folding (a :: (~>) s (f a)) (a :: (~>) a ((Const r :: Type -> Type) a)) (a :: s) :: (Const r :: Type -> Type) s where ...

Equations

Folding f g x = Case_6989586621679207593 f g x (Let6989586621679207589Scrutinee_6989586621679203569Sym3 f g x)

sFolding :: forall f r s a (t :: (~>) s (f a)) (t :: (~>) a ((Const r :: Type -> Type) a)) (t :: s). (SFoldable f, SMonoid r) => Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply FoldingSym0 t) t) t :: (Const r :: Type -> Type) s)

type Folded_ = FoldedSym0

The canonical Fold for any instance of Foldable.

Folded_
    :: Monoid r
    => Getting r (t a) a

type family Folded (a :: (~>) a ((Const r :: Type -> Type) a)) (a :: f a) :: (Const r :: Type -> Type) (f a) where ...

Equations

Folded f x = Case_6989586621679207576 f x (Let6989586621679207573Scrutinee_6989586621679203574Sym2 f x)

sFolded :: forall f r a (t :: (~>) a ((Const r :: Type -> Type) a)) (t :: f a). (SFoldable f, SMonoid r) => Sing t -> Sing t -> Sing (Apply (Apply FoldedSym0 t) t :: (Const r :: Type -> Type) (f a))

type Traverse_ = TraverseSym0

The canonical Traversal for any instance of Traversable.

Traverse_
    :: Applicative f
    => LensLike f (t a) (t b) a b

type family Traverse (arg :: a ~> f b) (arg1 :: t a) :: f (t b)

sTraverse :: (STraversable t, SApplicative f) => Sing t1 -> Sing t2 -> Sing (Apply (Apply (TraverseSym0 :: TyFun (a ~> f b) (t a ~> f (t b)) -> Type) t1) t2)

Cloning

type CloneTraversal_ l = CloneTraversalSym1 l

Clone a polymorphic traversal so it can be used as more than one type of thing (fold, traversal, getter, setter...).

CloneTraversal_
    :: Functor f
    => LensLike (Bazaar a b) s t a b
    -> LensLike f s t a b

Useful for writing a function that takes a traversal and uses it in more than one way; if you have it take an ATraversal, you can then use CloneTraversal_ to use it as a fold or traversal or anything else.

type family CloneTraversal (a :: (~>) ((~>) a (Bazaar a b b)) ((~>) s (Bazaar a b t))) (a :: (~>) a (f b)) (a :: s) :: f t where ...

Equations

CloneTraversal l f xs = Apply (Apply ($@#@$) (Apply UnBazaarSym0 f)) (Apply (Apply l (Apply (Apply (Apply Lambda_6989586621679207559Sym0 l) f) xs)) xs)

sCloneTraversal :: forall f a b s t (t :: (~>) ((~>) a (Bazaar a b b)) ((~>) s (Bazaar a b t))) (t :: (~>) a (f b)) (t :: s). SApplicative f => Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply CloneTraversalSym0 t) t) t :: f t)

Samples

Some sample lenses and traversals

* Tuple

type L1_ = L1Sym0

Lens into the first field of a tuple

type family L1 (a :: (~>) a (f b)) (a :: (a, c)) :: f (b, c) where ...

Equations

L1 f '(x, y) = Apply (Apply (<$>@#@$) (Apply (Apply (Apply Lambda_6989586621679207542Sym0 f) x) y)) (Apply f x)

sL1 :: forall f a c b (t :: (~>) a (f b)) (t :: (a, c)). SFunctor f => Sing t -> Sing t -> Sing (Apply (Apply L1Sym0 t) t :: f (b, c))

type L2_ = L2Sym0

Lens into the second field of a tuple

type family L2 (a :: (~>) b (f c)) (a :: (a, b)) :: f (a, c) where ...

Equations

L2 f '(x, y) = Apply (Apply (<$>@#@$) (Apply (Apply (Apply Lambda_6989586621679207527Sym0 f) x) y)) (Apply f y)

sL2 :: forall f a b c (t :: (~>) b (f c)) (t :: (a, b)). SFunctor f => Sing t -> Sing t -> Sing (Apply (Apply L2Sym0 t) t :: f (a, c))

List

data N

Peano nats, used for implementation of list index traversals in a termination-sane way.

Constructors

Z

S N

type SN = (Sing :: N -> Type)

type IxList_ i = IxListSym1 i

IxList i is a Traversal into the i-th item into a list. Defined in terms of N to allow for sane termination guaruntees.

IxList_
    :: Applicative f
    => N
    -> LensLike' f [a] a

type family IxList (a :: N) (a :: (~>) a (f a)) (a :: [a]) :: f [a] where ...

Equations

IxList _ _ '[] = Apply PureSym0 '[]

IxList Z f ((:) x xs) = Apply (Apply (<$>@#@$) (Apply (Apply (Apply Lambda_6989586621679207508Sym0 f) x) xs)) (Apply f x)

IxList (S i) f ((:) x xs) = Apply (Apply (<$>@#@$) (Apply (:@#@$) x)) (Apply (Apply (Apply IxListSym0 i) f) xs)

sIxList :: forall f a (t :: N) (t :: (~>) a (f a)) (t :: [a]). SApplicative f => Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply IxListSym0 t) t) t :: f [a])

Util

type (.@) f g = f .@#@$$$ g infixr 9

Shorter name for type-level function composition

data family Sing (a :: k) :: Type

The singleton kind-indexed data family.

Defunctionalization Symbols

data ASetterSym0 s6989586621679198796

data ASetterSym1 (s6989586621679198796 :: Type) t6989586621679198797

data ASetterSym2 (s6989586621679198796 :: Type) (t6989586621679198797 :: Type) a6989586621679198798

data ASetterSym3 (s6989586621679198796 :: Type) (t6989586621679198797 :: Type) (a6989586621679198798 :: Type) b6989586621679198799

type ASetterSym4 (s6989586621679198796 :: Type) (t6989586621679198797 :: Type) (a6989586621679198798 :: Type) (b6989586621679198799 :: Type) = ASetter s6989586621679198796 t6989586621679198797 a6989586621679198798 b6989586621679198799

data OverSym0 :: forall a6989586621679203447 b6989586621679203448 s6989586621679203445 t6989586621679203446. (~>) ((~>) ((~>) a6989586621679203447 (Identity b6989586621679203448)) ((~>) s6989586621679203445 (Identity t6989586621679203446))) ((~>) ((~>) a6989586621679203447 b6989586621679203448) ((~>) s6989586621679203445 t6989586621679203446))

data OverSym1 (a6989586621679207739 :: (~>) ((~>) a6989586621679203447 (Identity b6989586621679203448)) ((~>) s6989586621679203445 (Identity t6989586621679203446))) :: (~>) ((~>) a6989586621679203447 b6989586621679203448) ((~>) s6989586621679203445 t6989586621679203446)

data OverSym2 (a6989586621679207739 :: (~>) ((~>) a6989586621679203447 (Identity b6989586621679203448)) ((~>) s6989586621679203445 (Identity t6989586621679203446))) (a6989586621679207740 :: (~>) a6989586621679203447 b6989586621679203448) :: (~>) s6989586621679203445 t6989586621679203446

type OverSym3 (a6989586621679207739 :: (~>) ((~>) a6989586621679203447 (Identity b6989586621679203448)) ((~>) s6989586621679203445 (Identity t6989586621679203446))) (a6989586621679207740 :: (~>) a6989586621679203447 b6989586621679203448) (a6989586621679207741 :: s6989586621679203445) = Over a6989586621679207739 a6989586621679207740 a6989586621679207741

data SetSym0 :: forall a6989586621679203443 b6989586621679203444 s6989586621679203441 t6989586621679203442. (~>) ((~>) ((~>) a6989586621679203443 (Identity b6989586621679203444)) ((~>) s6989586621679203441 (Identity t6989586621679203442))) ((~>) b6989586621679203444 ((~>) s6989586621679203441 t6989586621679203442))

data SetSym1 (a6989586621679207756 :: (~>) ((~>) a6989586621679203443 (Identity b6989586621679203444)) ((~>) s6989586621679203441 (Identity t6989586621679203442))) :: (~>) b6989586621679203444 ((~>) s6989586621679203441 t6989586621679203442)

data SetSym2 (a6989586621679207756 :: (~>) ((~>) a6989586621679203443 (Identity b6989586621679203444)) ((~>) s6989586621679203441 (Identity t6989586621679203442))) (a6989586621679207757 :: b6989586621679203444) :: (~>) s6989586621679203441 t6989586621679203442

type SetSym3 (a6989586621679207756 :: (~>) ((~>) a6989586621679203443 (Identity b6989586621679203444)) ((~>) s6989586621679203441 (Identity t6989586621679203442))) (a6989586621679207757 :: b6989586621679203444) (a6989586621679207758 :: s6989586621679203441) = Set a6989586621679207756 a6989586621679207757 a6989586621679207758

data SetsSym0 :: forall a6989586621679203435 b6989586621679203436 s6989586621679203437 t6989586621679203438. (~>) ((~>) ((~>) a6989586621679203435 b6989586621679203436) ((~>) s6989586621679203437 t6989586621679203438)) ((~>) ((~>) a6989586621679203435 (Identity b6989586621679203436)) ((~>) s6989586621679203437 (Identity t6989586621679203438)))

data SetsSym1 (a6989586621679207698 :: (~>) ((~>) a6989586621679203435 b6989586621679203436) ((~>) s6989586621679203437 t6989586621679203438)) :: (~>) ((~>) a6989586621679203435 (Identity b6989586621679203436)) ((~>) s6989586621679203437 (Identity t6989586621679203438))

data SetsSym2 (a6989586621679207698 :: (~>) ((~>) a6989586621679203435 b6989586621679203436) ((~>) s6989586621679203437 t6989586621679203438)) (a6989586621679207699 :: (~>) a6989586621679203435 (Identity b6989586621679203436)) :: (~>) s6989586621679203437 (Identity t6989586621679203438)

type SetsSym3 (a6989586621679207698 :: (~>) ((~>) a6989586621679203435 b6989586621679203436) ((~>) s6989586621679203437 t6989586621679203438)) (a6989586621679207699 :: (~>) a6989586621679203435 (Identity b6989586621679203436)) (a6989586621679207700 :: s6989586621679203437) = Sets a6989586621679207698 a6989586621679207699 a6989586621679207700

data GettingSym0 r6989586621679198793

data GettingSym1 (r6989586621679198793 :: Type) s6989586621679198794

data GettingSym2 (r6989586621679198793 :: Type) (s6989586621679198794 :: Type) a6989586621679198795

type GettingSym3 (r6989586621679198793 :: Type) (s6989586621679198794 :: Type) (a6989586621679198795 :: Type) = Getting r6989586621679198793 s6989586621679198794 a6989586621679198795

data ViewSym0 :: forall a6989586621679203439 s6989586621679203440. (~>) ((~>) ((~>) a6989586621679203439 ((Const a6989586621679203439 :: Type -> Type) a6989586621679203439)) ((~>) s6989586621679203440 ((Const a6989586621679203439 :: Type -> Type) s6989586621679203440))) ((~>) s6989586621679203440 a6989586621679203439)

data ViewSym1 (a6989586621679207726 :: (~>) ((~>) a6989586621679203439 ((Const a6989586621679203439 :: Type -> Type) a6989586621679203439)) ((~>) s6989586621679203440 ((Const a6989586621679203439 :: Type -> Type) s6989586621679203440))) :: (~>) s6989586621679203440 a6989586621679203439

type ViewSym2 (a6989586621679207726 :: (~>) ((~>) a6989586621679203439 ((Const a6989586621679203439 :: Type -> Type) a6989586621679203439)) ((~>) s6989586621679203440 ((Const a6989586621679203439 :: Type -> Type) s6989586621679203440))) (a6989586621679207727 :: s6989586621679203440) = View a6989586621679207726 a6989586621679207727

data ToSym0 :: forall a6989586621679203433 r6989586621679203434 s6989586621679203432. (~>) ((~>) s6989586621679203432 a6989586621679203433) ((~>) ((~>) a6989586621679203433 ((Const r6989586621679203434 :: Type -> Type) a6989586621679203433)) ((~>) s6989586621679203432 ((Const r6989586621679203434 :: Type -> Type) s6989586621679203432)))

data ToSym1 (a6989586621679207681 :: (~>) s6989586621679203432 a6989586621679203433) :: forall r6989586621679203434. (~>) ((~>) a6989586621679203433 ((Const r6989586621679203434 :: Type -> Type) a6989586621679203433)) ((~>) s6989586621679203432 ((Const r6989586621679203434 :: Type -> Type) s6989586621679203432))

data ToSym2 (a6989586621679207681 :: (~>) s6989586621679203432 a6989586621679203433) (a6989586621679207682 :: (~>) a6989586621679203433 ((Const r6989586621679203434 :: Type -> Type) a6989586621679203433)) :: (~>) s6989586621679203432 ((Const r6989586621679203434 :: Type -> Type) s6989586621679203432)

type ToSym3 (a6989586621679207681 :: (~>) s6989586621679203432 a6989586621679203433) (a6989586621679207682 :: (~>) a6989586621679203433 ((Const r6989586621679203434 :: Type -> Type) a6989586621679203433)) (a6989586621679207683 :: s6989586621679203432) = To a6989586621679207681 a6989586621679207682 a6989586621679207683

data ToListOfSym0 :: forall a6989586621679203420 s6989586621679203421. (~>) ((~>) ((~>) a6989586621679203420 ((Const [a6989586621679203420] :: Type -> Type) a6989586621679203420)) ((~>) s6989586621679203421 ((Const [a6989586621679203420] :: Type -> Type) s6989586621679203421))) ((~>) s6989586621679203421 [a6989586621679203420])

data ToListOfSym1 (a6989586621679207623 :: (~>) ((~>) a6989586621679203420 ((Const [a6989586621679203420] :: Type -> Type) a6989586621679203420)) ((~>) s6989586621679203421 ((Const [a6989586621679203420] :: Type -> Type) s6989586621679203421))) :: (~>) s6989586621679203421 [a6989586621679203420]

type ToListOfSym2 (a6989586621679207623 :: (~>) ((~>) a6989586621679203420 ((Const [a6989586621679203420] :: Type -> Type) a6989586621679203420)) ((~>) s6989586621679203421 ((Const [a6989586621679203420] :: Type -> Type) s6989586621679203421))) (a6989586621679207624 :: s6989586621679203421) = ToListOf a6989586621679207623 a6989586621679207624

data LensLikeSym0 f6989586621679198803

data LensLikeSym1 (f6989586621679198803 :: k6989586621679198818 -> Type) s6989586621679198804

data LensLikeSym2 (f6989586621679198803 :: k6989586621679198818 -> Type) (s6989586621679198804 :: Type) t6989586621679198805

data LensLikeSym3 (f6989586621679198803 :: k6989586621679198818 -> Type) (s6989586621679198804 :: Type) (t6989586621679198805 :: k6989586621679198818) a6989586621679198806

data LensLikeSym4 (f6989586621679198803 :: k6989586621679198818 -> Type) (s6989586621679198804 :: Type) (t6989586621679198805 :: k6989586621679198818) (a6989586621679198806 :: Type) b6989586621679198807

type LensLikeSym5 (f6989586621679198803 :: k6989586621679198818 -> Type) (s6989586621679198804 :: Type) (t6989586621679198805 :: k6989586621679198818) (a6989586621679198806 :: Type) (b6989586621679198807 :: k6989586621679198818) = LensLike f6989586621679198803 s6989586621679198804 t6989586621679198805 a6989586621679198806 b6989586621679198807

data LensLike'Sym0 f6989586621679198800

data LensLike'Sym1 (f6989586621679198800 :: Type -> Type) s6989586621679198801

data LensLike'Sym2 (f6989586621679198800 :: Type -> Type) (s6989586621679198801 :: Type) a6989586621679198802

type LensLike'Sym3 (f6989586621679198800 :: Type -> Type) (s6989586621679198801 :: Type) (a6989586621679198802 :: Type) = LensLike' f6989586621679198800 s6989586621679198801 a6989586621679198802

data MkLensSym0 :: forall a6989586621679203429 b6989586621679203430 f6989586621679203427 s6989586621679203428 t6989586621679203431. (~>) ((~>) s6989586621679203428 a6989586621679203429) ((~>) ((~>) s6989586621679203428 ((~>) b6989586621679203430 t6989586621679203431)) ((~>) ((~>) a6989586621679203429 (f6989586621679203427 b6989586621679203430)) ((~>) s6989586621679203428 (f6989586621679203427 t6989586621679203431))))

data MkLensSym1 (a6989586621679207669 :: (~>) s6989586621679203428 a6989586621679203429) :: forall b6989586621679203430 f6989586621679203427 t6989586621679203431. (~>) ((~>) s6989586621679203428 ((~>) b6989586621679203430 t6989586621679203431)) ((~>) ((~>) a6989586621679203429 (f6989586621679203427 b6989586621679203430)) ((~>) s6989586621679203428 (f6989586621679203427 t6989586621679203431)))

data MkLensSym2 (a6989586621679207669 :: (~>) s6989586621679203428 a6989586621679203429) (a6989586621679207670 :: (~>) s6989586621679203428 ((~>) b6989586621679203430 t6989586621679203431)) :: forall f6989586621679203427. (~>) ((~>) a6989586621679203429 (f6989586621679203427 b6989586621679203430)) ((~>) s6989586621679203428 (f6989586621679203427 t6989586621679203431))

data MkLensSym3 (a6989586621679207669 :: (~>) s6989586621679203428 a6989586621679203429) (a6989586621679207670 :: (~>) s6989586621679203428 ((~>) b6989586621679203430 t6989586621679203431)) (a6989586621679207671 :: (~>) a6989586621679203429 (f6989586621679203427 b6989586621679203430)) :: (~>) s6989586621679203428 (f6989586621679203427 t6989586621679203431)

type MkLensSym4 (a6989586621679207669 :: (~>) s6989586621679203428 a6989586621679203429) (a6989586621679207670 :: (~>) s6989586621679203428 ((~>) b6989586621679203430 t6989586621679203431)) (a6989586621679207671 :: (~>) a6989586621679203429 (f6989586621679203427 b6989586621679203430)) (a6989586621679207672 :: s6989586621679203428) = MkLens a6989586621679207669 a6989586621679207670 a6989586621679207671 a6989586621679207672

data CloneLensSym0 :: forall a6989586621679203423 b6989586621679203424 f6989586621679203422 s6989586621679203425 t6989586621679203426. (~>) ((~>) ((~>) a6989586621679203423 (Context a6989586621679203423 b6989586621679203424 b6989586621679203424)) ((~>) s6989586621679203425 (Context a6989586621679203423 b6989586621679203424 t6989586621679203426))) ((~>) ((~>) a6989586621679203423 (f6989586621679203422 b6989586621679203424)) ((~>) s6989586621679203425 (f6989586621679203422 t6989586621679203426)))

data CloneLensSym1 (a6989586621679207643 :: (~>) ((~>) a6989586621679203423 (Context a6989586621679203423 b6989586621679203424 b6989586621679203424)) ((~>) s6989586621679203425 (Context a6989586621679203423 b6989586621679203424 t6989586621679203426))) :: forall f6989586621679203422. (~>) ((~>) a6989586621679203423 (f6989586621679203422 b6989586621679203424)) ((~>) s6989586621679203425 (f6989586621679203422 t6989586621679203426))

data CloneLensSym2 (a6989586621679207643 :: (~>) ((~>) a6989586621679203423 (Context a6989586621679203423 b6989586621679203424 b6989586621679203424)) ((~>) s6989586621679203425 (Context a6989586621679203423 b6989586621679203424 t6989586621679203426))) (a6989586621679207644 :: (~>) a6989586621679203423 (f6989586621679203422 b6989586621679203424)) :: (~>) s6989586621679203425 (f6989586621679203422 t6989586621679203426)

type CloneLensSym3 (a6989586621679207643 :: (~>) ((~>) a6989586621679203423 (Context a6989586621679203423 b6989586621679203424 b6989586621679203424)) ((~>) s6989586621679203425 (Context a6989586621679203423 b6989586621679203424 t6989586621679203426))) (a6989586621679207644 :: (~>) a6989586621679203423 (f6989586621679203422 b6989586621679203424)) (a6989586621679207645 :: s6989586621679203425) = CloneLens a6989586621679207643 a6989586621679207644 a6989586621679207645

data FoldingSym0 :: forall a6989586621679203415 f6989586621679203412 r6989586621679203413 s6989586621679203414. (~>) ((~>) s6989586621679203414 (f6989586621679203412 a6989586621679203415)) ((~>) ((~>) a6989586621679203415 ((Const r6989586621679203413 :: Type -> Type) a6989586621679203415)) ((~>) s6989586621679203414 ((Const r6989586621679203413 :: Type -> Type) s6989586621679203414)))

data FoldingSym1 (a6989586621679207580 :: (~>) s6989586621679203414 (f6989586621679203412 a6989586621679203415)) :: forall r6989586621679203413. (~>) ((~>) a6989586621679203415 ((Const r6989586621679203413 :: Type -> Type) a6989586621679203415)) ((~>) s6989586621679203414 ((Const r6989586621679203413 :: Type -> Type) s6989586621679203414))

data FoldingSym2 (a6989586621679207580 :: (~>) s6989586621679203414 (f6989586621679203412 a6989586621679203415)) (a6989586621679207581 :: (~>) a6989586621679203415 ((Const r6989586621679203413 :: Type -> Type) a6989586621679203415)) :: (~>) s6989586621679203414 ((Const r6989586621679203413 :: Type -> Type) s6989586621679203414)

type FoldingSym3 (a6989586621679207580 :: (~>) s6989586621679203414 (f6989586621679203412 a6989586621679203415)) (a6989586621679207581 :: (~>) a6989586621679203415 ((Const r6989586621679203413 :: Type -> Type) a6989586621679203415)) (a6989586621679207582 :: s6989586621679203414) = Folding a6989586621679207580 a6989586621679207581 a6989586621679207582

data FoldedSym0 :: forall a6989586621679203411 f6989586621679203409 r6989586621679203410. (~>) ((~>) a6989586621679203411 ((Const r6989586621679203410 :: Type -> Type) a6989586621679203411)) ((~>) (f6989586621679203409 a6989586621679203411) ((Const r6989586621679203410 :: Type -> Type) (f6989586621679203409 a6989586621679203411)))

data FoldedSym1 (a6989586621679207567 :: (~>) a6989586621679203411 ((Const r6989586621679203410 :: Type -> Type) a6989586621679203411)) :: forall f6989586621679203409. (~>) (f6989586621679203409 a6989586621679203411) ((Const r6989586621679203410 :: Type -> Type) (f6989586621679203409 a6989586621679203411))

type FoldedSym2 (a6989586621679207567 :: (~>) a6989586621679203411 ((Const r6989586621679203410 :: Type -> Type) a6989586621679203411)) (a6989586621679207568 :: f6989586621679203409 a6989586621679203411) = Folded a6989586621679207567 a6989586621679207568

data L1Sym0 :: forall a6989586621679203401 b6989586621679203403 c6989586621679203402 f6989586621679203400. (~>) ((~>) a6989586621679203401 (f6989586621679203400 b6989586621679203403)) ((~>) (a6989586621679203401, c6989586621679203402) (f6989586621679203400 (b6989586621679203403, c6989586621679203402)))

data L1Sym1 (a6989586621679207535 :: (~>) a6989586621679203401 (f6989586621679203400 b6989586621679203403)) :: forall c6989586621679203402. (~>) (a6989586621679203401, c6989586621679203402) (f6989586621679203400 (b6989586621679203403, c6989586621679203402))

type L1Sym2 (a6989586621679207535 :: (~>) a6989586621679203401 (f6989586621679203400 b6989586621679203403)) (a6989586621679207536 :: (a6989586621679203401, c6989586621679203402)) = L1 a6989586621679207535 a6989586621679207536

data L2Sym0 :: forall a6989586621679203397 b6989586621679203398 c6989586621679203399 f6989586621679203396. (~>) ((~>) b6989586621679203398 (f6989586621679203396 c6989586621679203399)) ((~>) (a6989586621679203397, b6989586621679203398) (f6989586621679203396 (a6989586621679203397, c6989586621679203399)))

data L2Sym1 (a6989586621679207520 :: (~>) b6989586621679203398 (f6989586621679203396 c6989586621679203399)) :: forall a6989586621679203397. (~>) (a6989586621679203397, b6989586621679203398) (f6989586621679203396 (a6989586621679203397, c6989586621679203399))

type L2Sym2 (a6989586621679207520 :: (~>) b6989586621679203398 (f6989586621679203396 c6989586621679203399)) (a6989586621679207521 :: (a6989586621679203397, b6989586621679203398)) = L2 a6989586621679207520 a6989586621679207521

type ZSym0 = Z

data SSym0 :: (~>) N N

type SSym1 (t6989586621679199042 :: N) = S t6989586621679199042

data IxListSym0 :: forall a6989586621679203395 f6989586621679203394. (~>) N ((~>) ((~>) a6989586621679203395 (f6989586621679203394 a6989586621679203395)) ((~>) [a6989586621679203395] (f6989586621679203394 [a6989586621679203395])))

data IxListSym1 (a6989586621679207499 :: N) :: forall a6989586621679203395 f6989586621679203394. (~>) ((~>) a6989586621679203395 (f6989586621679203394 a6989586621679203395)) ((~>) [a6989586621679203395] (f6989586621679203394 [a6989586621679203395]))

data IxListSym2 (a6989586621679207499 :: N) (a6989586621679207500 :: (~>) a6989586621679203395 (f6989586621679203394 a6989586621679203395)) :: (~>) [a6989586621679203395] (f6989586621679203394 [a6989586621679203395])

type IxListSym3 (a6989586621679207499 :: N) (a6989586621679207500 :: (~>) a6989586621679203395 (f6989586621679203394 a6989586621679203395)) (a6989586621679207501 :: [a6989586621679203395]) = IxList a6989586621679207499 a6989586621679207500 a6989586621679207501