Data.Type.Lens

Description

Lenses and optics for manipulating DataKind-based types, powered by singletons defunctionalization.

See Data.Type.Lens.Examples for example usage and syntax.

For the most part, you should be able to use them just like you'd use the functions from the lens or microlens libraries; just remember to capitalize names like Over and Set, since they are type families.

Note that the ways of "creating" a lens or optic (Sets_, Traverse_, To_, MkLens_, etc. are all suffixed with _ for convenience, to reserve the underscoreless identifiers for the fully applied type family as per singletons library convention.

"Value level" versions of these lenses and functions are not exported, because they would be more or less completely compatible with the same functions from microlens and lens packages. However, when the names of these functions differ from the names of their lens counterpart (mkLens and ixList), value-level versions are provided.

There are two main ways to define optics.

First, you can write them by hand using singletonsOnly:

 $(singletonsOnly [d|
   l1 :: Functor f => LensLike (a, c) (b, c) a b
   l1 f (x, y) = (x' -> (x', y)) $ f x

   l1Alt :: Functor f => LensLike (a, c) (b, c) a b
   l1Alt = mkLens fst ((_, y) x -> (x', y))

   getFirst :: Getting a (a, b) a
   getFirst = to fst
   |])

This creates the type families L1, L1Alt, and GetFirst; however, these aren't lenses, because they aren't partially applied. The lactual lenses are L1Sym0, L1AltSym0, and GetFirstSym0. As a convention, it is recommend to alias the actual lenses with an underscore suffix:

-- L1_       :: Functor f => LensLike f (a, c) (b, c) a b
type L1_       = L1Sym0

-- L1Alt_    :: Functor f => LensLike f (a, c) (b, c) a b
type L1Alt     = L1AltSym0

-- GetFirst_ :: Getting a (a, b) a
type GetFirst_ = GetFirstSym0

The number after the Sym is determined by how many arguments you need to apply to your function before you get to the actual lens. For example, IxList requires one argument (the index) to get to the actual traversal, so the definition in the library is:

type IxList_ i = IxListSym1 i

Second, you can write them directly at the type level using combinators like MkLens_ and To_:

type GetFirst_ = To_ FstSym0

(FstSym0 is the promotion of fst from Data.Singletons.Prelude.Tuple)

Documentation

type LensLike f s t a b = (a -> f b) -> s -> f t

The general shape of optics in this library. ("van Laarhoven")

For different levels of polymorphism on f, you get different types of optics:

If f can be any Functor, you have a Lens (see ALens)

If f is only Identity, you have a setter (see ASetter)

If f is only Const R for a specific R, you have a getter of R (see Getting)

If f can be Const r for any Monoid r, you have a Fold.

If f can be any Applicative, you have a Traversal (see ATraversal)

Normal lens libraries implement the constraints for lenses, folds, and traversals using RankN types, but we don't do that here to avoid working with RankN kinds.

type LensLike' f s a = LensLike f s s a a

A LensLike that does not change any types.

Setting

type ASetter s t a b = LensLike Identity s t a b

A settable "lens". Usable with Over (%~), constructable with To or any of the general lens constructors.

See LensLike for more information.

Using

Ways of consuming a setter.

type family Over (a :: (~>) ((~>) a (Identity b)) ((~>) s (Identity t))) (a :: (~>) a b) (a :: s) :: t where ...

Equations

Over l f x = Case_6989586621679215069 l f x (Let6989586621679215065Scrutinee_6989586621679210948Sym3 l f x)

type (%~) l f = OverSym2 l f infixr 4

Infix application of Over

sOver :: forall s t a b (t :: (~>) ((~>) a (Identity b)) ((~>) s (Identity t))) (t :: (~>) a b) (t :: s). Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply OverSym0 t) t) t :: t)

type family Set (a :: (~>) ((~>) a (Identity b)) ((~>) s (Identity t))) (a :: b) (a :: s) :: t where ...

Equations

Set l y a_6989586621679215079 = Apply (Apply (Apply OverSym0 l) (Apply (Apply (Apply Lambda_6989586621679215084Sym0 l) y) a_6989586621679215079)) a_6989586621679215079

type (.~) l x = SetSym2 l x infixr 4

Infix application of Set

sSet :: forall s t a b (t :: (~>) ((~>) a (Identity b)) ((~>) s (Identity t))) (t :: b) (t :: s). Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply SetSym0 t) t) t :: t)

Making

Ways of creating a setter-only.

type Sets_ f = SetsSym1 f

Create a Setter from a setting function.

Sets_ :: ((a ~> b) ~> (s ~> t)) -> ASetter s t a b

type family Sets (a :: (~>) ((~>) a b) ((~>) s t)) (a :: (~>) a (Identity b)) (a :: s) :: Identity t where ...

Equations

Sets f g a_6989586621679215021 = Apply (Apply (Apply (.@#@$) IdentitySym0) (Apply f (Apply (Apply (Apply Lambda_6989586621679215026Sym0 f) g) a_6989586621679215021))) a_6989586621679215021

sSets :: forall a b s t (t :: (~>) ((~>) a b) ((~>) s t)) (t :: (~>) a (Identity b)) (t :: s). Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply SetsSym0 t) t) t :: Identity t)

Getting

type Getting r s a = LensLike (Const r) s s a a

A retrieving "lens". If r is fixed to a type, it's a Getter for that type. If r is polymorphic over all Monoid, then it's a Fold over as.

As a Getter, usable with View (^.); as a Fold, usable with ToListOf (^..), Preview (^?), etc.

Normal lens libraries implement the constraints for folds using RankN types, but we don't do that here to avoid working with RankN kinds.

See LensLike for more information.

Using

Ways of consuming a getter

type family View (a :: (~>) ((~>) a ((Const a :: Type -> Type) a)) ((~>) s ((Const a :: Type -> Type) s))) (a :: s) :: a where ...

Equations

View l x = Case_6989586621679215052 l x (Let6989586621679215049Scrutinee_6989586621679210953Sym2 l x)

type (^.) x l = View l x infixl 8

Infix application of View

sView :: forall a s (t :: (~>) ((~>) a ((Const a :: Type -> Type) a)) ((~>) s ((Const a :: Type -> Type) s))) (t :: s). Sing t -> Sing t -> Sing (Apply (Apply ViewSym0 t) t :: a)

Making

Ways of creating a getter-only.

type To_ f = ToSym1 f

Create a Getter from a getting function.

To_ :: (a ~> b) -> Getting b a b

type family To (a :: (~>) s a) (a :: (~>) a ((Const r :: Type -> Type) a)) (a :: s) :: (Const r :: Type -> Type) s where ...

Equations

To f g x = Case_6989586621679215011 f g x (Let6989586621679215007Scrutinee_6989586621679210961Sym3 f g x)

sTo :: forall s a r (t :: (~>) s a) (t :: (~>) a ((Const r :: Type -> Type) a)) (t :: s). Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply ToSym0 t) t) t :: (Const r :: Type -> Type) s)

Lenses

type ALens s t a b = LensLike (Context a b) s t a b

If a function expects an ALens, it can be given any Lens (a LensLike f that works for any Functor f).

You can use an ALens as a normal lens by using CloneLens_.

Making

Ways of creating a lens

type MkLens_ f g = MkLensSym2 f g

Create a Lens from a setter and a getter.

MkLens_
    :: Functor f
    => (s ~> a)
    -> (s ~> b ~> t)
    -> LensLike f s t a b

type family MkLens (a :: (~>) s a) (a :: (~>) s ((~>) b t)) (a :: (~>) a (f b)) (a :: s) :: f t where ...

Equations

MkLens v s f x = Apply (Apply (<$>@#@$) (Apply s x)) (Apply f (Apply v x))

sMkLens :: forall f s a b t (t :: (~>) s a) (t :: (~>) s ((~>) b t)) (t :: (~>) a (f b)) (t :: s). SFunctor f => Sing t -> Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply (Apply MkLensSym0 t) t) t) t :: f t)

mkLens :: Functor f => (s -> a) -> (s -> b -> t) -> LensLike f s t a b

Cloning

type CloneLens_ l = CloneLensSym1 l

Clone a polymorphic lens so it can be used as more than one type of thing (getter or setter).

CloneLens_
    :: Functor f
    => LensLike (Context a b) s t a b
    -> LensLike f s t a b

Useful for writing a function that takes a lens and uses it in more than one way; if you have it take an ALens, you can then use CloneLens_ to use it as a getter or setter.

type family CloneLens (a :: (~>) ((~>) a (Context a b b)) ((~>) s (Context a b t))) (a :: (~>) a (f b)) (a :: s) :: f t where ...

Equations

CloneLens l f x = Case_6989586621679214917 l f x (Let6989586621679214905Scrutinee_6989586621679210996Sym3 l f x)

Traversals and Folds

type ATraversal s t a b = LensLike (Bazaar a b) s t a b

If a function expects an ATraversal, it can be given any Traversal (a LensLike f that works for any Applicative f).

You can use an ATraversal as a normal traversal by using CloneTraversal_.

Using

Ways of consuming traversals and folds

type family Preview (a :: (~>) ((~>) a ((Const (First a) :: Type -> Type) a)) ((~>) s ((Const (First a) :: Type -> Type) s))) (a :: s) :: Maybe a where ...

Equations

Preview l x = Case_6989586621679214961 l x (Let6989586621679214958Scrutinee_6989586621679210973Sym2 l x)

type (^?) x l = Preview l x infixl 8

Infix application of Preview

sPreview :: forall a s (t :: (~>) ((~>) a ((Const (First a) :: Type -> Type) a)) ((~>) s ((Const (First a) :: Type -> Type) s))) (t :: s). Sing t -> Sing t -> Sing (Apply (Apply PreviewSym0 t) t :: Maybe a)

type family ToListOf (a :: (~>) ((~>) a ((Const [a] :: Type -> Type) a)) ((~>) s ((Const [a] :: Type -> Type) s))) (a :: s) :: [a] where ...

Equations

ToListOf l x = Case_6989586621679214994 l x (Let6989586621679214984Scrutinee_6989586621679210966Sym2 l x)

type (^..) x l = ToListOf l x infixl 8

Infix application of ToListOf

sToListOf :: forall a s (t :: (~>) ((~>) a ((Const [a] :: Type -> Type) a)) ((~>) s ((Const [a] :: Type -> Type) s))) (t :: s). Sing t -> Sing t -> Sing (Apply (Apply ToListOfSym0 t) t :: [a])

type family UnsafePreview (a :: (~>) ((~>) a ((Const (First a) :: Type -> Type) a)) ((~>) s ((Const (First a) :: Type -> Type) s))) (a :: s) :: a where ...

Equations

UnsafePreview l x = Case_6989586621679214974 l x (Let6989586621679214971Scrutinee_6989586621679210980Sym2 l x)

type (^?!) x l = UnsafePreview l x infixl 8

Infix application of UnsafePreview

sUnsafePreview :: forall a s (t :: (~>) ((~>) a ((Const (First a) :: Type -> Type) a)) ((~>) s ((Const (First a) :: Type -> Type) s))) (t :: s). Sing t -> Sing t -> Sing (Apply (Apply UnsafePreviewSym0 t) t :: a)

Making

Ways of creating traversals and folds

type Folding_ f = FoldingSym1 f

Create a Fold from a "folding function":

Folding_
    :: (Foldable f, Monoid r)
    => (s ~> f a)
    -> Getting r s a

type family Folding (a :: (~>) s (f a)) (a :: (~>) a ((Const r :: Type -> Type) a)) (a :: s) :: (Const r :: Type -> Type) s where ...

Equations

Folding f g x = Case_6989586621679214948 f g x (Let6989586621679214944Scrutinee_6989586621679210986Sym3 f g x)

sFolding :: forall f r s a (t :: (~>) s (f a)) (t :: (~>) a ((Const r :: Type -> Type) a)) (t :: s). (SFoldable f, SMonoid r) => Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply FoldingSym0 t) t) t :: (Const r :: Type -> Type) s)

type Folded_ = FoldedSym0

The canonical Fold for any instance of Foldable.

Folded_
    :: Monoid r
    => Getting r (t a) a

type family Folded (a :: (~>) a ((Const r :: Type -> Type) a)) (a :: f a) :: (Const r :: Type -> Type) (f a) where ...

Equations

Folded f x = Case_6989586621679214931 f x (Let6989586621679214928Scrutinee_6989586621679210991Sym2 f x)

sFolded :: forall f r a (t :: (~>) a ((Const r :: Type -> Type) a)) (t :: f a). (SFoldable f, SMonoid r) => Sing t -> Sing t -> Sing (Apply (Apply FoldedSym0 t) t :: (Const r :: Type -> Type) (f a))

type Traverse_ = TraverseSym0

The canonical Traversal for any instance of Traversable.

Traverse_
    :: Applicative f
    => LensLike f (t a) (t b) a b

type family Traverse (arg :: a ~> f b) (arg1 :: t a) :: f (t b)

sTraverse :: (STraversable t, SApplicative f) => Sing t1 -> Sing t2 -> Sing (Apply (Apply (TraverseSym0 :: TyFun (a ~> f b) (t a ~> f (t b)) -> Type) t1) t2)

Cloning

type CloneTraversal_ l = CloneTraversalSym1 l

Clone a polymorphic traversal so it can be used as more than one type of thing (fold, traversal, getter, setter...).

CloneTraversal_
    :: Functor f
    => LensLike (Bazaar a b) s t a b
    -> LensLike f s t a b

Useful for writing a function that takes a traversal and uses it in more than one way; if you have it take an ATraversal, you can then use CloneTraversal_ to use it as a fold or traversal or anything else.

type family CloneTraversal (a :: (~>) ((~>) a (Bazaar a b b)) ((~>) s (Bazaar a b t))) (a :: (~>) a (f b)) (a :: s) :: f t where ...

Equations

CloneTraversal l f xs = Apply (Apply ($@#@$) (Apply UnBazaarSym0 f)) (Apply (Apply l (Apply (Apply (Apply Lambda_6989586621679214888Sym0 l) f) xs)) xs)

sCloneTraversal :: forall f a b s t (t :: (~>) ((~>) a (Bazaar a b b)) ((~>) s (Bazaar a b t))) (t :: (~>) a (f b)) (t :: s). SApplicative f => Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply CloneTraversalSym0 t) t) t :: f t)

Samples

Some sample lenses and traversals

* Tuple

type L1_ = L1Sym0

Lens into the first field of a tuple

type family L1 (a :: (~>) a (f b)) (a :: (a, c)) :: f (b, c) where ...

Equations

L1 f '(x, y) = Apply (Apply (<$>@#@$) (Apply (Apply (Apply Lambda_6989586621679214871Sym0 f) x) y)) (Apply f x)

sL1 :: forall f a c b (t :: (~>) a (f b)) (t :: (a, c)). SFunctor f => Sing t -> Sing t -> Sing (Apply (Apply L1Sym0 t) t :: f (b, c))

type L2_ = L2Sym0

Lens into the second field of a tuple

type family L2 (a :: (~>) b (f c)) (a :: (a, b)) :: f (a, c) where ...

Equations

L2 f '(x, y) = Apply (Apply (<$>@#@$) (Apply (Apply (Apply Lambda_6989586621679214856Sym0 f) x) y)) (Apply f y)

sL2 :: forall f a b c (t :: (~>) b (f c)) (t :: (a, b)). SFunctor f => Sing t -> Sing t -> Sing (Apply (Apply L2Sym0 t) t :: f (a, c))

List

data N

Peano nats, used for implementation of list index traversals in a termination-sane way.

Constructors

Z

S N

type SN = (Sing :: N -> Type)

type IxList_ i = IxListSym1 i

IxList i is a Traversal into the i-th item into a list. Defined in terms of N to allow for sane termination guaruntees.

IxList_
    :: Applicative f
    => N
    -> LensLike' f [a] a

type family IxList (a :: N) (a :: (~>) a (f a)) (a :: [a]) :: f [a] where ...

Equations

IxList _ _ '[] = Apply PureSym0 '[]

IxList Z f ((:) x xs) = Apply (Apply (<$>@#@$) (Apply (Apply (Apply Lambda_6989586621679239122Sym0 f) x) xs)) (Apply f x)

IxList (S i) f ((:) x xs) = Apply (Apply (<$>@#@$) (Apply (:@#@$) x)) (Apply (Apply (Apply IxListSym0 i) f) xs)

sIxList :: forall f a (t :: N) (t :: (~>) a (f a)) (t :: [a]). SApplicative f => Sing t -> Sing t -> Sing t -> Sing (Apply (Apply (Apply IxListSym0 t) t) t :: f [a])

ixList :: Applicative f => N -> LensLike' f [a] a

Util

type (.@) f g = f .@#@$$$ g infixr 9

Shorter name for type-level function composition

data family Sing (a :: k) :: Type

The singleton kind-indexed data family.

Defunctionalization Symbols

data ASetterSym0 s6989586621679201104

data ASetterSym1 (s6989586621679201104 :: Type) t6989586621679201105

data ASetterSym2 (s6989586621679201104 :: Type) (t6989586621679201105 :: Type) a6989586621679201106

data ASetterSym3 (s6989586621679201104 :: Type) (t6989586621679201105 :: Type) (a6989586621679201106 :: Type) b6989586621679201107

type ASetterSym4 (s6989586621679201104 :: Type) (t6989586621679201105 :: Type) (a6989586621679201106 :: Type) (b6989586621679201107 :: Type) = ASetter s6989586621679201104 t6989586621679201105 a6989586621679201106 b6989586621679201107

data OverSym0 :: forall a6989586621679210882 b6989586621679210883 s6989586621679210880 t6989586621679210881. (~>) ((~>) ((~>) a6989586621679210882 (Identity b6989586621679210883)) ((~>) s6989586621679210880 (Identity t6989586621679210881))) ((~>) ((~>) a6989586621679210882 b6989586621679210883) ((~>) s6989586621679210880 t6989586621679210881))

data OverSym1 (a6989586621679215056 :: (~>) ((~>) a6989586621679210882 (Identity b6989586621679210883)) ((~>) s6989586621679210880 (Identity t6989586621679210881))) :: (~>) ((~>) a6989586621679210882 b6989586621679210883) ((~>) s6989586621679210880 t6989586621679210881)

data OverSym2 (a6989586621679215056 :: (~>) ((~>) a6989586621679210882 (Identity b6989586621679210883)) ((~>) s6989586621679210880 (Identity t6989586621679210881))) (a6989586621679215057 :: (~>) a6989586621679210882 b6989586621679210883) :: (~>) s6989586621679210880 t6989586621679210881

type OverSym3 (a6989586621679215056 :: (~>) ((~>) a6989586621679210882 (Identity b6989586621679210883)) ((~>) s6989586621679210880 (Identity t6989586621679210881))) (a6989586621679215057 :: (~>) a6989586621679210882 b6989586621679210883) (a6989586621679215058 :: s6989586621679210880) = Over a6989586621679215056 a6989586621679215057 a6989586621679215058

data SetSym0 :: forall a6989586621679210878 b6989586621679210879 s6989586621679210876 t6989586621679210877. (~>) ((~>) ((~>) a6989586621679210878 (Identity b6989586621679210879)) ((~>) s6989586621679210876 (Identity t6989586621679210877))) ((~>) b6989586621679210879 ((~>) s6989586621679210876 t6989586621679210877))

data SetSym1 (a6989586621679215073 :: (~>) ((~>) a6989586621679210878 (Identity b6989586621679210879)) ((~>) s6989586621679210876 (Identity t6989586621679210877))) :: (~>) b6989586621679210879 ((~>) s6989586621679210876 t6989586621679210877)

data SetSym2 (a6989586621679215073 :: (~>) ((~>) a6989586621679210878 (Identity b6989586621679210879)) ((~>) s6989586621679210876 (Identity t6989586621679210877))) (a6989586621679215074 :: b6989586621679210879) :: (~>) s6989586621679210876 t6989586621679210877

type SetSym3 (a6989586621679215073 :: (~>) ((~>) a6989586621679210878 (Identity b6989586621679210879)) ((~>) s6989586621679210876 (Identity t6989586621679210877))) (a6989586621679215074 :: b6989586621679210879) (a6989586621679215075 :: s6989586621679210876) = Set a6989586621679215073 a6989586621679215074 a6989586621679215075

data SetsSym0 :: forall a6989586621679210870 b6989586621679210871 s6989586621679210872 t6989586621679210873. (~>) ((~>) ((~>) a6989586621679210870 b6989586621679210871) ((~>) s6989586621679210872 t6989586621679210873)) ((~>) ((~>) a6989586621679210870 (Identity b6989586621679210871)) ((~>) s6989586621679210872 (Identity t6989586621679210873)))

data SetsSym1 (a6989586621679215015 :: (~>) ((~>) a6989586621679210870 b6989586621679210871) ((~>) s6989586621679210872 t6989586621679210873)) :: (~>) ((~>) a6989586621679210870 (Identity b6989586621679210871)) ((~>) s6989586621679210872 (Identity t6989586621679210873))

data SetsSym2 (a6989586621679215015 :: (~>) ((~>) a6989586621679210870 b6989586621679210871) ((~>) s6989586621679210872 t6989586621679210873)) (a6989586621679215016 :: (~>) a6989586621679210870 (Identity b6989586621679210871)) :: (~>) s6989586621679210872 (Identity t6989586621679210873)

type SetsSym3 (a6989586621679215015 :: (~>) ((~>) a6989586621679210870 b6989586621679210871) ((~>) s6989586621679210872 t6989586621679210873)) (a6989586621679215016 :: (~>) a6989586621679210870 (Identity b6989586621679210871)) (a6989586621679215017 :: s6989586621679210872) = Sets a6989586621679215015 a6989586621679215016 a6989586621679215017

data GettingSym0 r6989586621679201101

data GettingSym1 (r6989586621679201101 :: Type) s6989586621679201102

data GettingSym2 (r6989586621679201101 :: Type) (s6989586621679201102 :: Type) a6989586621679201103

type GettingSym3 (r6989586621679201101 :: Type) (s6989586621679201102 :: Type) (a6989586621679201103 :: Type) = Getting r6989586621679201101 s6989586621679201102 a6989586621679201103

data ViewSym0 :: forall a6989586621679210874 s6989586621679210875. (~>) ((~>) ((~>) a6989586621679210874 ((Const a6989586621679210874 :: Type -> Type) a6989586621679210874)) ((~>) s6989586621679210875 ((Const a6989586621679210874 :: Type -> Type) s6989586621679210875))) ((~>) s6989586621679210875 a6989586621679210874)

data ViewSym1 (a6989586621679215043 :: (~>) ((~>) a6989586621679210874 ((Const a6989586621679210874 :: Type -> Type) a6989586621679210874)) ((~>) s6989586621679210875 ((Const a6989586621679210874 :: Type -> Type) s6989586621679210875))) :: (~>) s6989586621679210875 a6989586621679210874

type ViewSym2 (a6989586621679215043 :: (~>) ((~>) a6989586621679210874 ((Const a6989586621679210874 :: Type -> Type) a6989586621679210874)) ((~>) s6989586621679210875 ((Const a6989586621679210874 :: Type -> Type) s6989586621679210875))) (a6989586621679215044 :: s6989586621679210875) = View a6989586621679215043 a6989586621679215044

data ToSym0 :: forall a6989586621679210868 r6989586621679210869 s6989586621679210867. (~>) ((~>) s6989586621679210867 a6989586621679210868) ((~>) ((~>) a6989586621679210868 ((Const r6989586621679210869 :: Type -> Type) a6989586621679210868)) ((~>) s6989586621679210867 ((Const r6989586621679210869 :: Type -> Type) s6989586621679210867)))

data ToSym1 (a6989586621679214998 :: (~>) s6989586621679210867 a6989586621679210868) :: forall r6989586621679210869. (~>) ((~>) a6989586621679210868 ((Const r6989586621679210869 :: Type -> Type) a6989586621679210868)) ((~>) s6989586621679210867 ((Const r6989586621679210869 :: Type -> Type) s6989586621679210867))

data ToSym2 (a6989586621679214998 :: (~>) s6989586621679210867 a6989586621679210868) (a6989586621679214999 :: (~>) a6989586621679210868 ((Const r6989586621679210869 :: Type -> Type) a6989586621679210868)) :: (~>) s6989586621679210867 ((Const r6989586621679210869 :: Type -> Type) s6989586621679210867)

type ToSym3 (a6989586621679214998 :: (~>) s6989586621679210867 a6989586621679210868) (a6989586621679214999 :: (~>) a6989586621679210868 ((Const r6989586621679210869 :: Type -> Type) a6989586621679210868)) (a6989586621679215000 :: s6989586621679210867) = To a6989586621679214998 a6989586621679214999 a6989586621679215000

data ToListOfSym0 :: forall a6989586621679210865 s6989586621679210866. (~>) ((~>) ((~>) a6989586621679210865 ((Const [a6989586621679210865] :: Type -> Type) a6989586621679210865)) ((~>) s6989586621679210866 ((Const [a6989586621679210865] :: Type -> Type) s6989586621679210866))) ((~>) s6989586621679210866 [a6989586621679210865])

data ToListOfSym1 (a6989586621679214978 :: (~>) ((~>) a6989586621679210865 ((Const [a6989586621679210865] :: Type -> Type) a6989586621679210865)) ((~>) s6989586621679210866 ((Const [a6989586621679210865] :: Type -> Type) s6989586621679210866))) :: (~>) s6989586621679210866 [a6989586621679210865]

type ToListOfSym2 (a6989586621679214978 :: (~>) ((~>) a6989586621679210865 ((Const [a6989586621679210865] :: Type -> Type) a6989586621679210865)) ((~>) s6989586621679210866 ((Const [a6989586621679210865] :: Type -> Type) s6989586621679210866))) (a6989586621679214979 :: s6989586621679210866) = ToListOf a6989586621679214978 a6989586621679214979

data LensLikeSym0 f6989586621679201111

data LensLikeSym1 (f6989586621679201111 :: k6989586621679201126 -> Type) s6989586621679201112

data LensLikeSym2 (f6989586621679201111 :: k6989586621679201126 -> Type) (s6989586621679201112 :: Type) t6989586621679201113

data LensLikeSym3 (f6989586621679201111 :: k6989586621679201126 -> Type) (s6989586621679201112 :: Type) (t6989586621679201113 :: k6989586621679201126) a6989586621679201114

data LensLikeSym4 (f6989586621679201111 :: k6989586621679201126 -> Type) (s6989586621679201112 :: Type) (t6989586621679201113 :: k6989586621679201126) (a6989586621679201114 :: Type) b6989586621679201115

type LensLikeSym5 (f6989586621679201111 :: k6989586621679201126 -> Type) (s6989586621679201112 :: Type) (t6989586621679201113 :: k6989586621679201126) (a6989586621679201114 :: Type) (b6989586621679201115 :: k6989586621679201126) = LensLike f6989586621679201111 s6989586621679201112 t6989586621679201113 a6989586621679201114 b6989586621679201115

data LensLike'Sym0 f6989586621679201108

data LensLike'Sym1 (f6989586621679201108 :: Type -> Type) s6989586621679201109

data LensLike'Sym2 (f6989586621679201108 :: Type -> Type) (s6989586621679201109 :: Type) a6989586621679201110

type LensLike'Sym3 (f6989586621679201108 :: Type -> Type) (s6989586621679201109 :: Type) (a6989586621679201110 :: Type) = LensLike' f6989586621679201108 s6989586621679201109 a6989586621679201110

data MkLensSym0 :: forall a6989586621679239055 b6989586621679239056 f6989586621679239053 s6989586621679239054 t6989586621679239057. (~>) ((~>) s6989586621679239054 a6989586621679239055) ((~>) ((~>) s6989586621679239054 ((~>) b6989586621679239056 t6989586621679239057)) ((~>) ((~>) a6989586621679239055 (f6989586621679239053 b6989586621679239056)) ((~>) s6989586621679239054 (f6989586621679239053 t6989586621679239057))))

data MkLensSym1 (a6989586621679239134 :: (~>) s6989586621679239054 a6989586621679239055) :: forall b6989586621679239056 f6989586621679239053 t6989586621679239057. (~>) ((~>) s6989586621679239054 ((~>) b6989586621679239056 t6989586621679239057)) ((~>) ((~>) a6989586621679239055 (f6989586621679239053 b6989586621679239056)) ((~>) s6989586621679239054 (f6989586621679239053 t6989586621679239057)))

data MkLensSym2 (a6989586621679239134 :: (~>) s6989586621679239054 a6989586621679239055) (a6989586621679239135 :: (~>) s6989586621679239054 ((~>) b6989586621679239056 t6989586621679239057)) :: forall f6989586621679239053. (~>) ((~>) a6989586621679239055 (f6989586621679239053 b6989586621679239056)) ((~>) s6989586621679239054 (f6989586621679239053 t6989586621679239057))

data MkLensSym3 (a6989586621679239134 :: (~>) s6989586621679239054 a6989586621679239055) (a6989586621679239135 :: (~>) s6989586621679239054 ((~>) b6989586621679239056 t6989586621679239057)) (a6989586621679239136 :: (~>) a6989586621679239055 (f6989586621679239053 b6989586621679239056)) :: (~>) s6989586621679239054 (f6989586621679239053 t6989586621679239057)

type MkLensSym4 (a6989586621679239134 :: (~>) s6989586621679239054 a6989586621679239055) (a6989586621679239135 :: (~>) s6989586621679239054 ((~>) b6989586621679239056 t6989586621679239057)) (a6989586621679239136 :: (~>) a6989586621679239055 (f6989586621679239053 b6989586621679239056)) (a6989586621679239137 :: s6989586621679239054) = MkLens a6989586621679239134 a6989586621679239135 a6989586621679239136 a6989586621679239137

data CloneLensSym0 :: forall a6989586621679210850 b6989586621679210851 f6989586621679210849 s6989586621679210852 t6989586621679210853. (~>) ((~>) ((~>) a6989586621679210850 (Context a6989586621679210850 b6989586621679210851 b6989586621679210851)) ((~>) s6989586621679210852 (Context a6989586621679210850 b6989586621679210851 t6989586621679210853))) ((~>) ((~>) a6989586621679210850 (f6989586621679210849 b6989586621679210851)) ((~>) s6989586621679210852 (f6989586621679210849 t6989586621679210853)))

data CloneLensSym1 (a6989586621679214896 :: (~>) ((~>) a6989586621679210850 (Context a6989586621679210850 b6989586621679210851 b6989586621679210851)) ((~>) s6989586621679210852 (Context a6989586621679210850 b6989586621679210851 t6989586621679210853))) :: forall f6989586621679210849. (~>) ((~>) a6989586621679210850 (f6989586621679210849 b6989586621679210851)) ((~>) s6989586621679210852 (f6989586621679210849 t6989586621679210853))

data CloneLensSym2 (a6989586621679214896 :: (~>) ((~>) a6989586621679210850 (Context a6989586621679210850 b6989586621679210851 b6989586621679210851)) ((~>) s6989586621679210852 (Context a6989586621679210850 b6989586621679210851 t6989586621679210853))) (a6989586621679214897 :: (~>) a6989586621679210850 (f6989586621679210849 b6989586621679210851)) :: (~>) s6989586621679210852 (f6989586621679210849 t6989586621679210853)

type CloneLensSym3 (a6989586621679214896 :: (~>) ((~>) a6989586621679210850 (Context a6989586621679210850 b6989586621679210851 b6989586621679210851)) ((~>) s6989586621679210852 (Context a6989586621679210850 b6989586621679210851 t6989586621679210853))) (a6989586621679214897 :: (~>) a6989586621679210850 (f6989586621679210849 b6989586621679210851)) (a6989586621679214898 :: s6989586621679210852) = CloneLens a6989586621679214896 a6989586621679214897 a6989586621679214898

data FoldingSym0 :: forall a6989586621679210860 f6989586621679210857 r6989586621679210858 s6989586621679210859. (~>) ((~>) s6989586621679210859 (f6989586621679210857 a6989586621679210860)) ((~>) ((~>) a6989586621679210860 ((Const r6989586621679210858 :: Type -> Type) a6989586621679210860)) ((~>) s6989586621679210859 ((Const r6989586621679210858 :: Type -> Type) s6989586621679210859)))

data FoldingSym1 (a6989586621679214935 :: (~>) s6989586621679210859 (f6989586621679210857 a6989586621679210860)) :: forall r6989586621679210858. (~>) ((~>) a6989586621679210860 ((Const r6989586621679210858 :: Type -> Type) a6989586621679210860)) ((~>) s6989586621679210859 ((Const r6989586621679210858 :: Type -> Type) s6989586621679210859))

data FoldingSym2 (a6989586621679214935 :: (~>) s6989586621679210859 (f6989586621679210857 a6989586621679210860)) (a6989586621679214936 :: (~>) a6989586621679210860 ((Const r6989586621679210858 :: Type -> Type) a6989586621679210860)) :: (~>) s6989586621679210859 ((Const r6989586621679210858 :: Type -> Type) s6989586621679210859)

type FoldingSym3 (a6989586621679214935 :: (~>) s6989586621679210859 (f6989586621679210857 a6989586621679210860)) (a6989586621679214936 :: (~>) a6989586621679210860 ((Const r6989586621679210858 :: Type -> Type) a6989586621679210860)) (a6989586621679214937 :: s6989586621679210859) = Folding a6989586621679214935 a6989586621679214936 a6989586621679214937

data FoldedSym0 :: forall a6989586621679210856 f6989586621679210854 r6989586621679210855. (~>) ((~>) a6989586621679210856 ((Const r6989586621679210855 :: Type -> Type) a6989586621679210856)) ((~>) (f6989586621679210854 a6989586621679210856) ((Const r6989586621679210855 :: Type -> Type) (f6989586621679210854 a6989586621679210856)))

data FoldedSym1 (a6989586621679214922 :: (~>) a6989586621679210856 ((Const r6989586621679210855 :: Type -> Type) a6989586621679210856)) :: forall f6989586621679210854. (~>) (f6989586621679210854 a6989586621679210856) ((Const r6989586621679210855 :: Type -> Type) (f6989586621679210854 a6989586621679210856))

type FoldedSym2 (a6989586621679214922 :: (~>) a6989586621679210856 ((Const r6989586621679210855 :: Type -> Type) a6989586621679210856)) (a6989586621679214923 :: f6989586621679210854 a6989586621679210856) = Folded a6989586621679214922 a6989586621679214923

data L1Sym0 :: forall a6989586621679210841 b6989586621679210843 c6989586621679210842 f6989586621679210840. (~>) ((~>) a6989586621679210841 (f6989586621679210840 b6989586621679210843)) ((~>) (a6989586621679210841, c6989586621679210842) (f6989586621679210840 (b6989586621679210843, c6989586621679210842)))

data L1Sym1 (a6989586621679214864 :: (~>) a6989586621679210841 (f6989586621679210840 b6989586621679210843)) :: forall c6989586621679210842. (~>) (a6989586621679210841, c6989586621679210842) (f6989586621679210840 (b6989586621679210843, c6989586621679210842))

type L1Sym2 (a6989586621679214864 :: (~>) a6989586621679210841 (f6989586621679210840 b6989586621679210843)) (a6989586621679214865 :: (a6989586621679210841, c6989586621679210842)) = L1 a6989586621679214864 a6989586621679214865

data L2Sym0 :: forall a6989586621679210837 b6989586621679210838 c6989586621679210839 f6989586621679210836. (~>) ((~>) b6989586621679210838 (f6989586621679210836 c6989586621679210839)) ((~>) (a6989586621679210837, b6989586621679210838) (f6989586621679210836 (a6989586621679210837, c6989586621679210839)))

data L2Sym1 (a6989586621679214849 :: (~>) b6989586621679210838 (f6989586621679210836 c6989586621679210839)) :: forall a6989586621679210837. (~>) (a6989586621679210837, b6989586621679210838) (f6989586621679210836 (a6989586621679210837, c6989586621679210839))

type L2Sym2 (a6989586621679214849 :: (~>) b6989586621679210838 (f6989586621679210836 c6989586621679210839)) (a6989586621679214850 :: (a6989586621679210837, b6989586621679210838)) = L2 a6989586621679214849 a6989586621679214850

type ZSym0 = Z

data SSym0 :: (~>) N N

type SSym1 (t6989586621679202515 :: N) = S t6989586621679202515

data IxListSym0 :: forall a6989586621679239052 f6989586621679239051. (~>) N ((~>) ((~>) a6989586621679239052 (f6989586621679239051 a6989586621679239052)) ((~>) [a6989586621679239052] (f6989586621679239051 [a6989586621679239052])))

data IxListSym1 (a6989586621679239113 :: N) :: forall a6989586621679239052 f6989586621679239051. (~>) ((~>) a6989586621679239052 (f6989586621679239051 a6989586621679239052)) ((~>) [a6989586621679239052] (f6989586621679239051 [a6989586621679239052]))

data IxListSym2 (a6989586621679239113 :: N) (a6989586621679239114 :: (~>) a6989586621679239052 (f6989586621679239051 a6989586621679239052)) :: (~>) [a6989586621679239052] (f6989586621679239051 [a6989586621679239052])

type IxListSym3 (a6989586621679239113 :: N) (a6989586621679239114 :: (~>) a6989586621679239052 (f6989586621679239051 a6989586621679239052)) (a6989586621679239115 :: [a6989586621679239052]) = IxList a6989586621679239113 a6989586621679239114 a6989586621679239115